// api-client/models/EbookSignUpRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bright Website API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface EbookSignUpRequest
 */
export interface EbookSignUpRequest {
    /**
     * 
     * @type {string}
     * @memberof EbookSignUpRequest
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof EbookSignUpRequest
     */
    email: string;
    /**
     * 
     * @type {boolean}
     * @memberof EbookSignUpRequest
     */
    consent: boolean;
}

/**
 * Check if a given object implements the EbookSignUpRequest interface.
 */
export function instanceOfEbookSignUpRequest(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "email" in value;
    isInstance = isInstance && "consent" in value;

    return isInstance;
}

export function EbookSignUpRequestFromJSON(json: any): EbookSignUpRequest {
    return EbookSignUpRequestFromJSONTyped(json, false);
}

export function EbookSignUpRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): EbookSignUpRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'email': json['email'],
        'consent': json['consent'],
    };
}

export function EbookSignUpRequestToJSON(value?: EbookSignUpRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'email': value.email,
        'consent': value.consent,
    };
}